middleware: extract bindAll into utils file
